# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IIRP V1",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/diagnostics/policy": {
      "get": {
        "summary": "Get Policy",
        "operationId": "get_policy_api_v1_diagnostics_policy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyView"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Policy",
        "operationId": "patch_policy_api_v1_diagnostics_policy_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/diagnostics/collections": {
      "post": {
        "summary": "Collection",
        "operationId": "collection_api_v1_diagnostics_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/actions": {
      "post": {
        "summary": "Action",
        "operationId": "action_api_v1_jobs__job_id__actions_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/feed": {
      "get": {
        "summary": "Feed",
        "operationId": "feed_api_v1_feed_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Session Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Cursor"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "all",
              "title": "Type"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "transaction",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/search": {
      "get": {
        "summary": "Search",
        "operationId": "search_api_v1_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenericOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{digest}": {
      "get": {
        "summary": "Source",
        "operationId": "source_api_v1_sources__digest__get",
        "parameters": [
          {
            "name": "digest",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Digest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "pause",
              "resume",
              "cancel",
              "retry"
            ],
            "title": "Action"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "ActionRequest"
      },
...
      "CollectionRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "fixture_check",
              "market_probe",
              "sec_probe"
            ],
            "title": "Kind"
          },
          "ticker": {
            "type": "string",
            "pattern": "^[A-Za-z0-9^][A-Za-z0-9.^=-]{0,14}$",
            "title": "Ticker",
            "default": "AAPL"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "CollectionRequest"
      },
      "CollectionResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "reused": {
            "type": "boolean",
            "title": "Reused"
          },
          "job": {
            "$ref": "#/components/schemas/JobView"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "reused",
          "job"
        ],
        "title": "CollectionResponse"
      },
...
      "FeedOutput": {
        "properties": {
          "groups": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Groups"
          },
          "order": {
            "type": "string",
            "title": "Order",
            "default": "transaction"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "as_of": {
            "type": "string",
            "title": "As Of"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "total_groups": {
            "type": "integer",
            "title": "Total Groups"
          },
          "data_status": {
            "type": "string",
            "title": "Data Status"
          },
          "coverage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Coverage"
          },
          "new_count": {
            "type": "integer",
            "title": "New Count",
            "default": 0
          },
          "pending_filings": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Pending Filings"
          },
          "pending_summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Pending Summary"
          }
        },
        "type": "object",
        "required": [
          "groups",
          "session_id",
          "as_of",
          "next_cursor",
          "total_groups",
          "data_status",
          "coverage"
        ],
        "title": "FeedOutput"
      },
...
      "GenericOutput": {
        "properties": {
          "items": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Items"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "title": "GenericOutput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trigger": {
            "type": "string",
            "title": "Trigger"
          },
          "progress_done": {
            "type": "integer",
            "title": "Progress Done"
          },
          "progress_total": {
            "type": "integer",
            "title": "Progress Total"
          },
          "checkpoint": {
            "additionalProperties": true,
            "type": "object",
            "title": "Checkpoint"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "requested_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested Action"
          },
          "control_version": {
            "type": "integer",
            "title": "Control Version"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "control_notice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Control Notice"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "title",
          "status",
          "trigger",
          "progress_done",
          "progress_total",
          "checkpoint",
          "attempts",
          "error",
          "created_at",
          "updated_at",
          "started_at",
          "finished_at",
          "requested_action",
          "control_version",
          "result"
        ],
        "title": "JobView"
      },
...
      "PolicyRequest": {
        "properties": {
          "sec_enabled": {
            "type": "boolean",
            "title": "Sec Enabled"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sec_enabled"
        ],
        "title": "PolicyRequest"
      },
      "PolicyView": {
        "properties": {
          "sec_enabled": {
            "type": "boolean",
            "title": "Sec Enabled"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "next_run_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Run At"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          }
        },
        "type": "object",
        "required": [
          "sec_enabled",
          "version",
          "updated_at",
          "next_run_at",
          "scope"
        ],
        "title": "PolicyView"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}